+++ test/RuleEngine/AccessControl/RuleEngineAccessControl.sol
@@ -91,7 +91,9 @@ contract RuleEngineTest is Test, HelperContract {
     function testCannotAttackerOperateOnTransfer() public {
         // Act
         vm.prank(ATTACKER);
-        vm.expectRevert(ERC3643ComplianceModuleInvariantStorage.RuleEngine_ERC3643Compliance_UnauthorizedCaller.selector);
+        vm.expectRevert(
+            ERC3643ComplianceModuleInvariantStorage.RuleEngine_ERC3643Compliance_UnauthorizedCaller.selector
+        );
         ruleEngineMock.transferred(address(0), ADDRESS1, ADDRESS2, 10);
     }